//更新内容 1、添加了可以修改【保存游戏数据】的速度 2、添加了可以调节基础战斗速度 3、

+++ server/src/main/java/com/eztv/mud/command/commands/faction/DestroyFaction.java
@@ -11,6 +11,11 @@ import com.eztv.mud.constant.Enum;
 
 import static com.eztv.mud.GameUtil.*;
 
+/**
+ * 作者: hhx QQ1025334900
+ * 时间: 2020-08-05 10:21
+ * 功能: 解散门派
+ **/
 public class DestroyFaction extends BaseCommand {
     public DestroyFaction(Client client, Msg msg, String key) {
         super(client, msg, key);
@@ -18,14 +23,25 @@ public class DestroyFaction extends BaseCommand {
 
     @Override
     public void execute() {
+        if(getPlayer().getFaction_position()!=5){//不是帮主
+            //发送帮派解散成功消息
+            String sendStr = getPropByFile("faction","faction_destroy_fail_grant",
+                    getClient().getPlayer().getName(),"");
+            Chat chat = new Chat();
+            chat.setContent(sendStr);
+            chat.setMsgType(Enum.chat.系统);
+            GameUtil.sendToSelf(getClient(), msgBuild(Enum.messageType.chat, "公聊", object2JsonStr(chat), ""));
+            return;
+        }
         int flag = DataBase.getInstance().init().createSQL("delete from t_faction where id = '"+getClient().getPlayer().getFaction()+"'")
                 .update();
         if(flag>0){
             FactionCache.remove(getClient().getPlayer().getFaction());//从缓存中移除
             getClient().getPlayer().setFaction(0);
+            getClient().getPlayer().setFaction_position(0);
             getClient().getPlayer().getDataBaseHandler().savePlayer(getClient().getPlayer());
             //发送帮派解散成功消息
-            String sendStr = getProp("faction_destroy_success",
+            String sendStr = getPropByFile("faction","faction_destroy_success",
             getClient().getPlayer().getName(),"");
             Chat chat = new Chat();
             chat.setContent(sendStr);
